server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "channels>=4.3.2",
    "channels-redis>=4.3.0",
    "django>=6.0.2",
    "django-cors-headers>=4.9.0",
    "django-environ>=0.13.0",
    "django-modeltranslation>=0.20.2",
    "django-treenode>=0.23.3",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "drf-spectacular[sidecar]>=0.29.0",
    "pendulum>=3.2.0",
    "pillow>=12.1.1",
    "psycopg>=3.3.3",
    "python-dotenv>=1.2.2",
    "redis>=7.3.0",
    "rich>=14.3.3",
    "uvicorn>=0.41.0",
    "websockets>=16.0",
]
